Move hooks to the hooks folder
