+++ src/hooks/useScrollReached/index.ts
@@ -1,6 +1,6 @@
 import {useState, useCallback} from 'react';
 import {NativeSyntheticEvent, NativeScrollEvent} from 'react-native';
-import {Constants} from '../helpers';
+import {Constants} from '../../helpers';
 
 export type Props = {
   /**
